small refactor (make it inner class)
