+++ CHANGELOG.md
@@ -14,4 +14,4 @@
 
 ### Bug Fixes
 - Fix `--version` command output [#8960](https://github.com/Consensys/teku/issues/8960)
-- Fix issue introduced in 24.12.1 with peer stability when the upperbound is set to a high number
+- Fix issue (introduced in `24.12.1`) with peer stability when the upperbound is set to a high number
